pyproject.toml
[tool.poetry]
name = "example_template"
version = "1.0.0"
description = "This is a template example for creating internal Python Packages"
authors = ["Overjet <info@overjet.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
torch = "^2.0.0"
torchvision = "^0.15.1"
torchmetrics = "^0.11.4"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
colorama = "^0.4.6"
tqdm = "^4.65.0"
mlxtend = "^0.21.0"
hydra-core = "^1.3.2"
tomli = "1.2.3"
timm = "^0.6.13"
wandb = "^0.14.0"
neptune = "^1.1.1"
typing-extensions = "4.3.0"
typed-ast = "^1.5.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py38"]
exclude = '''
/(
    \.git
  | \.venv
  | \.direnv
  | _build
  | build
  | dist
  | \.env
  | \.vscode
  | \.cache
  | github
  | __pycache__
  | .cloudignore
)/
'''

[tool.isort]
py_version = "38"
line_length = 100
lines_after_imports = 2
lines_between_sections = 1
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = "true"
combine_as_imports = "true"
use_parentheses = "true"
skip = ["./docs", ".direnv"]
skip_glob = ["*/__init__.py"]
skip_gitignore = "true"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"

[tool.pytest.ini_options]
minversion = "6.1"
addopts = "--cov"
testpaths = [
    "tests_unit",
]
